+++ src/infrastructure/three/sceneSetup/index.ts
@@ -2,7 +2,7 @@ export { addHelpers } from './helpers';
 export { addDynamicGrid } from './dynamicGrid';
 export { snapObjectToFloor } from './floorSnap';
 export { fitShadowCameraToObject } from './shadowRig';
-export { bakeContactShadow, setContactShadowMode } from './contactShadow';
+export { bakeContactShadow, resetContactShadow, setContactShadowMode } from './contactShadow';
 export { wrapInUnitsScaleGroup } from './unitsScale';
 export { addLighting } from './lighting';
 export { createGradientBackground } from './gradientBackground';